requirements.txt
# Dev requirements
flake8
pytest

# Code requirements
numpy
tensorflow
Pillow
clearml

# API requirements
fastapi
pydantic
uvicorn
python-multipart

